feat(tests): enhance shader test files and debug output handling
- Updated `Cargo.toml` to include the "debug" feature for `lpvm-native-fa`.
- Modified `fa3-out-not-read.glsl` to change output vector type from `vec3` to `ivec3` and adjusted expected output.
- Added a new line in `native-rv32-iadd.glsl` for clarity.
- Improved error handling in `run_detail.rs` to prevent duplicate error messages.
- Introduced `lpir_module` method in `LpvmModule` trait and its implementations for `NativeEmuModule` and `NativeFaModule`.
- Enhanced debug output in `debug.rs` to provide clearer compile-time information.

These changes improve the robustness of shader testing and enhance the clarity of debug information during shader execution.

+++ lp-shader/lps-filetests/filetests/debug/fa3-out-not-read.glsl
@@ -1,16 +1,16 @@
 // test run
 
-void set_components(out vec3 result) {
-    result.x = 1.0;
-    result.y = 2.0;
-    result.z = 3.0;
+void set_components(out ivec3 result) {
+    result.x = 1;
+    result.y = 2;
+    result.z = 3;
 }
 
 float test_edge_out_modify_components() {
     // Can modify individual components without reading whole vector
-    vec3 v;
+    ivec3 v;
     set_components(v);
     return v.x + v.y + v.z;
 }
 
-// run: test_edge_out_modify_components() ~= 6.0
+// run: test_edge_out_modify_components() == 6
